+++ python/instrumentation/openinference-instrumentation-agno/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.1.27](https://github.com/Arize-ai/openinference/compare/python-openinference-instrumentation-agno-v0.1.26...python-openinference-instrumentation-agno-v0.1.27) (2026-01-27)
+
+
+### Bug Fixes
+
+* ensure tool error spans are properly ended and exported ([#2654](https://github.com/Arize-ai/openinference/issues/2654)) ([e58f177](https://github.com/Arize-ai/openinference/commit/e58f17794f3bde2998dc9f08e1d3307ebd8f96f4))
+
 ## [0.1.26](https://github.com/Arize-ai/openinference/compare/python-openinference-instrumentation-agno-v0.1.25...python-openinference-instrumentation-agno-v0.1.26) (2026-01-20)