+++ python/openinference-instrumentation/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.1.43](https://github.com/Arize-ai/openinference/compare/python-openinference-instrumentation-v0.1.42...python-openinference-instrumentation-v0.1.43) (2026-01-26)
+
+
+### Features
+
+* add Open Agent Spec Instrumentor ([#2560](https://github.com/Arize-ai/openinference/issues/2560)) ([48a8722](https://github.com/Arize-ai/openinference/commit/48a87228767089416bbe608a8ff782923c9db5b5))
+
 ## [0.1.42](https://github.com/Arize-ai/openinference/compare/python-openinference-instrumentation-v0.1.41...python-openinference-instrumentation-v0.1.42) (2025-11-05)